receipt-analyzer/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
structlog>=23.1.0
pytesseract>=0.3.10
pytest>=7.4.0
httpx>=0.24.1